backend/requirements.txt
# ===== LLM Providers =====
# Gemini
langchain-google-genai==4.2.0

# Hugging Face (v4 Stable Stack)
transformers==4.57.5
torch==2.9.1
huggingface-hub==0.35.0
langchain-huggingface==1.2.0

# ===== Core Framework & Agents =====
langchain==1.2.4
langchain-community==0.4.1
langgraph==1.0.6

# ===== API & Backend =====
flask==3.1.2
flask-cors==6.0.2
flasgger==0.9.7.1
python-dotenv==1.0.1

# ===== Utilities & Text Processing =====
requests==2.32.5
pydantic==2.12.5
aiohttp==3.13.3
marshmallow==3.21.3
sentencepiece==0.2.1
accelerate==1.12.0
psutil==6.1.1